Linting fixes
